+++ CHANGELOG.md
@@ -1,11 +1,17 @@
+## master
+
+### enhancements
+
+### bug fixes
+  * Specifying the button size (e.g. `btn-lg`) should still generate the appropriate button class
+  (`btn-default` or `btn-primary`). [@lowjoel](https://github.com/lowjoel)
+
 ## 1.1.0
 
 ### enhancements
   * There is no longer a need to retain a Simple Form initialiser, as well as a Simple 
   Form-Bootstrap initialiser. Just call `SimpleForm::setup` and the Bootstrap defaults would be 
-  automatically applied. [@lowjoel](https://github.com/lowjoel) 
-  [@rhodrid](https://github.com/rhodrid)
+  automatically applied. [@lowjoel](https://github.com/lowjoel)
   * Started writing specs and running them on Travis!  
 
 ### bug fixes
-  *